+++ tests/__init__.py
@@ -1,0 +1,3 @@
+"""
+Test package for content-aggregator-shared.
+"""